Remove execFile dep in terminal instance, unify 2 path prep functions
