+++ src/vs/workbench/contrib/terminal/browser/terminalPanel.ts
@@ -6,7 +6,6 @@
 import * as dom from 'vs/base/browser/dom';
 import * as nls from 'vs/nls';
 import * as platform from 'vs/base/common/platform';
-import * as terminalEnvironment from 'vs/workbench/contrib/terminal/common/terminalEnvironment';
 import { Action, IAction } from 'vs/base/common/actions';
 import { IActionItem, Separator } from 'vs/base/browser/ui/actionbar/actionbar';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
@@ -281,7 +280,7 @@ export class TerminalPanel extends Panel {
 				event.stopPropagation();
 			}
 		}));
-		this._register(dom.addDisposableListener(this._parentDomElement, dom.EventType.DROP, (e: DragEvent) => {
+		this._register(dom.addDisposableListener(this._parentDomElement, dom.EventType.DROP, async (e: DragEvent) => {
 			if (e.target === this._parentDomElement || dom.isAncestor(e.target as HTMLElement, this._parentDomElement)) {
 				if (!e.dataTransfer) {
 					return;
@@ -303,7 +302,9 @@ export class TerminalPanel extends Panel {
 
 				const terminal = this._terminalService.getActiveInstance();
 				if (terminal) {
-					terminal.sendText(terminalEnvironment.preparePathForTerminal(path), false);
+					return terminal.preparePathForTerminalAsync(path).then(preparedPath => {
+						terminal.sendText(preparedPath, false);
+					});
 				}
 			}
 		}));
